Refactor: 홈, 이벤트 레이아웃 코드 통일
