+++ Pyonsnal-Color/Pyonsnal-Color/Common/Interface/CommonProduct.swift
@@ -1,0 +1,19 @@
+//
+//  CommonProduct.swift
+//  Pyonsnal-Color
+//
+//  Created by 조소정 on 2023/09/03.
+//
+
+import UIKit
+
+// MARK: - Interface
+enum CommonProduct {
+    enum Size {
+        static let titleNavigationBarHeight: CGFloat = 48
+        static let storeHeight: CGFloat = 44
+        static let storeCollectionViewSeparatorHeight: CGFloat = 1
+        static let filterMargin: UIEdgeInsets = .init(top: 12, left: 15, bottom: 12, right: 15)
+        static let filterHeight: CGFloat = 56
+    }
+}
