+++ Pyonsnal-Color/Pyonsnal-Color/EventHome/EventHomeViewHolder.swift
@@ -1,47 +1,41 @@
 //
-//  ProductHomeViewHolder.swift
+//  EventHomeViewHolder.swift
 //  Pyonsnal-Color
 //
-//  Created by 김인호 on 2023/06/21.
+//  Created by 조소정 on 2023/09/03.
 //
 
 import UIKit
 import SnapKit
 
 // MARK: - UI Component
-extension ProductHomeViewController {
+extension EventHomeViewController {
     
     final class ViewHolder: ViewHolderable {
         let containerScrollView: UIScrollView = {
             let scrollView = UIScrollView()
-            scrollView.showsVerticalScrollIndicator = false
-            scrollView.showsHorizontalScrollIndicator = false
             scrollView.bounces = false
             return scrollView
         }()
         
         private let contentView: UIView = {
-            let view = UIView(frame: .zero)
+            let view = UIView()
             view.backgroundColor = .white
             return view
         }()
         
         let titleNavigationView = TitleNavigationView()
-        
-        let collectionStackView: UIStackView = {
-            let stackView = UIStackView()
-            stackView.axis = .vertical
-            return stackView
-        }()
+
         
         let collectionView: UICollectionView = {
-            let layout = UICollectionViewFlowLayout()
-            let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
+            let flowLayout = UICollectionViewFlowLayout()
+            let collectionView = UICollectionView(frame: .zero,
+                                                  collectionViewLayout: flowLayout)
             collectionView.backgroundColor = .clear
             collectionView.bounces = false
             return collectionView
         }()
-        
+                
         let storeCollectionViewSeparator: UIView = {
             let view: UIView = .init(frame: .zero)
             view.backgroundColor = .gray200
@@ -51,13 +45,18 @@ extension ProductHomeViewController {
         let filterCollectionView: UICollectionView = {
             let layout = UICollectionViewFlowLayout()
             let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
-            collectionView.backgroundColor = .clear
             collectionView.bounces = false
+            collectionView.backgroundColor = .clear
             collectionView.layoutMargins = CommonProduct.Size.filterMargin
             return collectionView
         }()
         
-        let productHomePageViewController: ProductHomePageViewController = .init(
+        private let contentPageView: UIView = {
+            let view = UIView()
+            return view
+        }()
+        
+        lazy var pageViewController = EventHomePageViewController(
             transitionStyle: .scroll,
             navigationOrientation: .horizontal
         )
@@ -65,45 +64,40 @@ extension ProductHomeViewController {
         func place(in view: UIView) {
             view.addSubview(containerScrollView)
             containerScrollView.addSubview(contentView)
-            
             contentView.addSubview(titleNavigationView)
             contentView.addSubview(storeCollectionViewSeparator)
-            contentView.addSubview(collectionStackView)
-            contentView.addSubview(productHomePageViewController.view)
-            
-            collectionStackView.addArrangedSubview(collectionView)
-            collectionStackView.addArrangedSubview(filterCollectionView)
+            contentView.addSubview(collectionView)
+            contentView.addSubview(filterCollectionView)
+            contentView.addSubview(contentPageView)
+            contentPageView.addSubview(pageViewController.view)
         }
         
         func configureConstraints(for view: UIView) {
-            containerScrollView.snp.makeConstraints { make in
-                make.edges.equalTo(view.safeAreaLayoutGuide)
+            containerScrollView.snp.makeConstraints {
+                $0.edges.equalTo(view.safeAreaLayoutGuide)
             }
             
-            contentView.snp.makeConstraints { make in
-                make.edges.equalToSuperview()
-                make.width.equalToSuperview()
-                make.height.equalToSuperview()
+            contentView.snp.makeConstraints {
+                $0.edges.equalToSuperview()
+                $0.width.equalToSuperview()
+                $0.height.equalToSuperview()
             }
             
-            titleNavigationView.snp.makeConstraints { make in
-                make.height.equalTo(CommonProduct.Size.titleNavigationBarHeight)
-                make.leading.trailing.top.equalToSuperview()
+            titleNavigationView.snp.makeConstraints {
+                $0.height.equalTo(CommonProduct.Size.titleNavigationBarHeight)
+                $0.top.leading.trailing.equalToSuperview()
             }
             
-            collectionStackView.snp.makeConstraints {
-                $0.leading.trailing.equalToSuperview()
+            collectionView.snp.makeConstraints {
                 $0.top.equalTo(titleNavigationView.snp.bottom)
-            }
-            
-            collectionView.snp.makeConstraints { make in
-                make.leading.trailing.equalToSuperview().inset(.spacing16)
-                make.height.equalTo(CommonProduct.Size.storeHeight)
+                $0.leading.equalToSuperview().offset(.spacing16)
+                $0.trailing.equalToSuperview().inset(.spacing16)
+                $0.height.equalTo(CommonProduct.Size.storeHeight)
             }
             
             storeCollectionViewSeparator.snp.makeConstraints { make in
                 make.height.equalTo(CommonProduct.Size.storeCollectionViewSeparatorHeight)
-                make.bottom.equalTo(collectionView.snp.bottom)
+                make.top.equalTo(collectionView.snp.bottom).inset(CommonProduct.Size.storeCollectionViewSeparatorHeight)
                 make.leading.trailing.equalToSuperview()
             }
             
@@ -113,10 +107,14 @@ extension ProductHomeViewController {
                 $0.leading.trailing.equalToSuperview()
             }
             
-            productHomePageViewController.view.snp.makeConstraints { make in
-                make.top.equalTo(collectionStackView.snp.bottom)
-                make.leading.trailing.equalToSuperview()
-                make.bottom.equalTo(view.safeAreaLayoutGuide)
+            contentPageView.snp.makeConstraints {
+                $0.top.equalTo(filterCollectionView.snp.bottom)
+                $0.leading.trailing.bottom.equalToSuperview()
+            }
+
+            pageViewController.view.snp.makeConstraints {
+                $0.top.leading.trailing.equalToSuperview()
+                $0.bottom.equalTo(view.safeAreaLayoutGuide)
             }
         }
     }
